I can only go one way. I've not got a reverse gear

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -9,7 +9,7 @@
  */
 void print_rev(char *s)
 {
-	long long int i;
+	long int i;
 
 	i = 0;
